% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialDiallel.R
\name{PartialDiallel}
\alias{PartialDiallel}
\title{Analysis for Partial Diallel}
\usage{
PartialDiallel(y, Rep, Cross1, Cross2, data)
}
\arguments{
\item{y}{Numeric Response Vector}

\item{Rep}{Replicate as factor}

\item{Cross1}{Cross 1 as factor}

\item{Cross2}{Cross 2 as factor}

\item{data}{A \code{data.frame}}
}
\value{
Means Means

ANOVA Analysis of Variance (ANOVA) table

Genetic.Components Genetic Components

General General

Specific Specific
}
\description{
Analysis of Partial Diallel
}
\examples{

data(PartialDiallelData)
fm1 <-
  PartialDiallel(
      y      = y
    , Rep    = Rep
    , Cross1 = Cross1
    , Cross2 = Cross2
    , data   = PartialDiallelData
  )

fm1
}
\seealso{
\code{\link{PartialDiallelData}}
 , \code{\link{Griffing}}
 , \code{\link{Hayman}}
 , \code{\link{GriffingData1}}
 , \code{\link{GriffingData2}}
 , \code{\link{GriffingData3}}
 , \code{\link{GriffingData4}}
}
\author{
\enumerate{
\item Pedro A. M. Barbosa (\email{pedro.barbosa@usp.br})
\item Muhammad Yaseen (\email{myaseen208@gmail.com})
 }
}
