% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PartialDiallelData.R
\docType{data}
\name{PartialDiallelData}
\alias{PartialDiallelData}
\title{Data for Partial Diallel Analysis}
\usage{
data(PartialDiallelData)
}
\description{
Data for Partial Diallel Analysis
}
\details{
\itemize{
       \item Cross1 Cross 1
       \item Cross2 Cross 2
       \item Rep Replicate
       \item Yield Yield Response
       }
}
\examples{
data(PartialDiallelData)
}
\seealso{
\code{\link{PartialDiallel}}
 , \code{\link{Griffing}}
 , \code{\link{Hayman}}
 , \code{\link{GriffingData1}}
 , \code{\link{GriffingData2}}
 , \code{\link{GriffingData3}}
 , \code{\link{GriffingData4}}
}
\keyword{datasets}
